% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_name_parsing.R
\name{decodeComparisonConditionMapping}
\alias{decodeComparisonConditionMapping}
\title{This function orchestrates the conversion of encoded strings
in the comparison condition mapping to the original conditions}
\usage{
decodeComparisonConditionMapping(conditionComparisonMapping, conditions_dict)
}
\description{
This function orchestrates the conversion of encoded strings
in the comparison condition mapping to the original conditions
}
