% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{des_tmt_PXD019880}
\alias{des_tmt_PXD019880}
\title{TMT Data processed with MaxQuant}
\format{
an experiment desing (des) data.table
}
\usage{
des_tmt_PXD019880
}
\description{
TMT Data processed with MaxQuant
}
\details{
The data can be used to run the LFQProcessing limma tmt pipeline \code{tmt_transformer}.

The processed data was published as part of http://doi.org/10.1021/acs.jproteome.0c00209

The original raw data was published as part of https://pubs.acs.org/doi/10.1021/pr401278j
}
\references{
PRIDE Project PXD019880 https://www.ebi.ac.uk/pride/archive/projects/PXD019880
}
\keyword{datasets}
