% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_exp_type.R
\name{detect_exp_type}
\alias{detect_exp_type}
\title{Detect Experiment Type}
\usage{
detect_exp_type(folder)
}
\arguments{
\item{folder}{A maxquant txt output folder}
}
\value{
A string describing the type of experiment
}
\description{
Detect Experiment Type
}
