% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_lfq_her2_targetted_therapy_tables}
\alias{example_lfq_her2_targetted_therapy_tables}
\title{HER2 Targetted Therapy processed MaxQuant Data}
\format{
List containing the maxquant output.txt folder as data.tables.
}
\usage{
example_lfq_her2_targetted_therapy_tables
}
\description{
HER2 Targetted Therapy processed MaxQuant Data
}
\details{
The data can be used to run the LFQProcessing limma lfq pipeline \code{lfq_transformer}.

The raw data has been processed in MaxQuant as described here: https://www.biorxiv.org/content/10.1101/2021.03.03.433806v1
}
\references{
PRIDE Project PXD002057 https://www.ebi.ac.uk/pride/archive/projects/PXD002057
}
\keyword{datasets}
