% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report_names_mapping.R
\name{get_names_qc_lfq_all}
\alias{get_names_qc_lfq_all}
\title{MaxQuant LFQ full QC plots names}
\usage{
get_names_qc_lfq_all()
}
\description{
QC report names of the plots created for a MaxQuant LFQ analysis including all files.
}
