% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report_names_mapping.R
\name{get_names_qc_lfq_protein_only}
\alias{get_names_qc_lfq_protein_only}
\title{MaxQuant FQ protein-only QC plots names}
\usage{
get_names_qc_lfq_protein_only()
}
\description{
QC report names of the plots created for a MaxQuant LFQ protein only analysis.
}
