% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report_names_mapping.R
\name{get_names_qc_tmt_protein_only}
\alias{get_names_qc_tmt_protein_only}
\title{MaxQuant TMT protein-only QC plots names}
\usage{
get_names_qc_tmt_protein_only(evidence = FALSE)
}
\arguments{
\item{evidence}{bool. TRUE to return QC also available when the evidence.txt file is uploaded.}
}
\description{
QC report names of the plots created for a MaxQuant TMT protein-only analysis.
If \code{evidence} file is provided, extra plots are produced.
}
