% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_viz.R
\name{get_protein_viz}
\alias{get_protein_viz}
\title{Print a Protein Visualization JSON file.}
\usage{
get_protein_viz(prot, prot_int, output_folder, conditionComparisonMapping)
}
\arguments{
\item{prot}{a processed proteinGroups table}

\item{prot_int}{a protein intensities table}

\item{output_folder}{a folder to write the json file into}
}
\value{
Writes a json object to the output folder
}
\description{
Print a Protein Visualization JSON file.
}
