% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_name_parsing.R
\name{getOriginalFromSafe}
\alias{getOriginalFromSafe}
\title{This function is a decoder utility that uses the conditions dict}
\usage{
getOriginalFromSafe(safe_word, conditions_dict)
}
\description{
This function is a decoder utility that uses the conditions dict
}
