% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_lfq.R
\name{impute_lfq}
\alias{impute_lfq}
\title{impute LFQ data}
\usage{
impute_lfq(
  myQuantDT,
  id_type,
  int_type,
  f_imputeStDev = imputeStDev,
  f_imputePosition = imputePosition
)
}
\arguments{
\item{myQuantDT}{quantification data}

\item{id_type}{the id column of myQuantDT that indicates}

\item{int_type}{the column of myQuantDT to be imputed}

\item{f_imputePosition}{The Position parameter for MNAR Imputation}

\item{f_imputStDev}{The Standard Deviation parameter for MNAR Imputation}
}
\value{
quantification data with missing values imputed
}
\description{
This function does MNAR imputation
}
