% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_stats_fun.R
\name{limma_stats_fun}
\alias{limma_stats_fun}
\title{This function performs the differential expression analysis with limma including all pairwise comparisons
using the condition provided}
\usage{
limma_stats_fun(
  ID_type,
  int_type,
  condition_col_name,
  run_id_col_name,
  rep_col_name,
  funDT,
  pairwise.comp = NULL,
  all.comparisons = TRUE,
  fix_distr = FALSE
)
}
\description{
This function performs the differential expression analysis with limma including all pairwise comparisons
using the condition provided
}
