% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrichment_table.R
\name{phospho_enrichment_quant}
\alias{phospho_enrichment_quant}
\title{Run Phosphosite Quantification on each level of Maxquant output}
\usage{
phospho_enrichment_quant(upload_folder, des, output_folder)
}
\arguments{
\item{upload_folder}{A maxquant txt output folder.}

\item{des}{An experiment design table}

\item{output_folder}{A folder to write output files to.}
}
\value{
Writes output files to the output folder
}
\description{
Run Phosphosite Quantification on each level of Maxquant output
}
\examples{
phospho_enrichment_quant(upload_folder, expdes, output_folder)
}
