% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_quant_runner.R
\name{protein_quant_runner}
\alias{protein_quant_runner}
\title{Run Pairwise Protein Quantification and Quality Control on Maxquant output}
\usage{
protein_quant_runner(upload_folder, output_folder)
}
\arguments{
\item{output_folder}{An output folder to store produced files.}

\item{folder}{A maxquant txt output folder.}
}
\value{
A string describing the type of experiment
}
\description{
Run Pairwise Protein Quantification and Quality Control on Maxquant output
}
