% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_separate_qcs.R
\name{render_separate_qcs}
\alias{render_separate_qcs}
\title{Render separate QC reports}
\usage{
render_separate_qcs(
  experiment_type,
  protein_only,
  evidence_exists,
  output_folder
)
}
\description{
Render separate QC reports
}
