% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Write Output of quantitative analysis}
\usage{
write_output(data, output_folder, file_name)
}
\arguments{
\item{data}{a data.table to write}

\item{output_folder}{An output folder to store produced files.}
}
\value{
Write the table to disk in the output folder
}
\description{
Write Output of quantitative analysis
}
