% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_viz.R
\name{writeReplicateData}
\alias{writeReplicateData}
\title{Write Protein counts and intensities json}
\usage{
writeReplicateData(prot_int, prot, outputFolder)
}
\arguments{
\item{prot_int}{a protein intensities table}

\item{prot}{a processed proteinGroups table}

\item{output_folder}{a folder to write the json file into
Columns required: \code{ProteinId}, \code{GeneName}, \code{Description}, \code{log2NInt}, \code{Condition},
\code{Replicate}, \code{Imputed}, \code{ProteinQValue}, \code{ProteinScore},  \code{FastaHeaders}.}
}
\description{
Write Protein counts and intensities json
}
