% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_quant_runner.R
\name{protein_quant_runner}
\alias{protein_quant_runner}
\title{Run Pairwise Protein Quantification and Quality Control on Maxquant output}
\usage{
protein_quant_runner(
  upload_folder,
  output_folder,
  protein_only = FALSE,
  write_qc = TRUE
)
}
\arguments{
\item{output_folder}{An output folder to store produced files.}

\item{protein_only}{Boolean, TRUE means LFQ will only process protein level data}

\item{folder}{A maxquant txt output folder.}
}
\value{
A string describing the type of experiment
}
\description{
Run Pairwise Protein Quantification and Quality Control on Maxquant output
}
